Create DanteBubbleInsertionSelectionSortBenchmarking
Simple Implementation of Bubble, insertion, and selection sort. Had the issue where bubble sort was incredibly fast until I realized it was sorting an array that was already sorted by insertion sort. So, I had to add code to make clones of the arrays and then feed the clones into each algorithm.